Extracted ImagePath from Image.kt.
