+++ backend/widgets/src/main/kotlin/br/com/zup/beagle/widget/ui/Image.kt
@@ -30,34 +30,3 @@ import br.com.zup.beagle.widget.core.ImageContentMode
 data class Image(val path: Bind<ImagePath>, val mode: ImageContentMode? = null) : Widget() {
     constructor(path: ImagePath, mode: ImageContentMode? = null) : this(valueOf(path), mode)
 }
-
-/**
- * Define the source of image data to populate the image view.
- *
- * @param url the URL from which the image can be obtained
- * @param placeholder the image that will used as placeholder when set a remote image.
- * */
-sealed class ImagePath(val url: String?, val placeholder: Local? = null) {
-    /**
-     * Define an image whose data is local to the client app.
-     *
-     * @param webUrl reference the path to an image in your web app files.
-     * @param mobileId reference an image natively in your mobile app local styles file.
-     * */
-    class Local internal constructor(webUrl: String?, val mobileId: String?) : ImagePath(webUrl) {
-        companion object {
-            fun both(webUrl: String, mobileId: String) = Local(webUrl, mobileId)
-            fun justMobile(mobileId: String) = Local(null, mobileId)
-            fun justWeb(webUrl: String) = Local(webUrl, null)
-        }
-    }
-
-    /**
-     * Define an image whose data needs to be downloaded from a source external to the client app.
-     *
-     * @param remoteUrl reference the path where the image should be fetched from.
-     * @param placeholder reference an image natively in your mobile app local styles file to be used as placeholder.
-     * */
-    class Remote(remoteUrl: String, placeholder: Local? = null) : ImagePath(remoteUrl, placeholder)
-}
-
